refactor: folder structure
move common domain folders into /common folder
