+++ src/domains/users/common/api/index.tsx
@@ -9,8 +9,8 @@ import router from "next/router";
 import { deleteAuthHeader, setAuthHeader } from "../utils";
 import { routerPush } from "@Utils/router";
 import { PasswordLoginCredentials } from "../types";
-import { LoginFormState } from "../Login";
-import { SignUpFormState } from "../ProviderSignUp/hooks";
+import { LoginFormState } from "@Users/Login/LoginPage";
+import { SignUpFormState } from "@Users/CustomerSignUp/hooks";
 
 export const signIn = async (formValues: LoginFormState) => {
   const signInResponse = nextAuthSignIn("credentials", {
